requirements.txt
locust
pyyaml
requests
